creating a separate file for the Options struct
